+++ R/Iterator.R
@@ -1,10 +1,10 @@
 #' Making an Irregular Sequence Iterator
 #'
 #' Create an Iterator object, where the user defines a sequence and a
-#' set of initial values, and then calls 'yield_next' to generate the
+#' set of initial values, and then calls `yield_next()` to generate the
 #' next element of the sequence. `Iterator`s are R environments, which means
-#' they are modified in place. To make a copy of an Iterator that can be modified
-#' separately, see [clone()].
+#' they are modified in place, even when passed as arguments to functions.
+#' To make a copy of an Iterator that can be modified separately, see [clone()].
 #'
 #' @param result R expression to run each time 'yield_next' is called
 #' @param initial named list or vector; declare and initialize every variable that appears in 'result'
@@ -13,8 +13,7 @@
 #' @return An environment object of S3 type Iterator
 #'
 #' @note The expression to be evaluated can include constant values not defined in
-#' `$initial` as long as they are defined in the search path of the environment
-#' where `yield_*` is called from. These values will not vary from iteration to
+#' `$initial` as long as they are defined globally. These values will not vary from iteration to
 #' iteration. Notably, `$initial` is a list, not an environment. As such, the environment
 #' in which the `Iterator` is created does not matter at all for how the expression is
 #' evaluated—all that matters is the environment where `yield_*` is called from.
@@ -29,7 +28,7 @@
 #' seq <- yield_more(collatz, 30)
 #'
 #' # using objects defined outside `$initial`:
-#' # Note that `n` in `$initial` overrides `n` in the search path
+#' # Note that `n` in `$initial` overrides the global `n`
 #' m <- 100
 #' n <- 10
 #' expr <- "out <- n + m"
@@ -54,12 +53,12 @@ Iterator <- function(result,
                       yield) {
   yield <- rlang::enexpr(yield)
   if (length(result) != 1L || !is.character(result)) {
-    rlang::abort("`result` must be a character string of length 1")
+    rlang::abort("`result` must be a single character string")
   }
   result <- rlang::parse_exprs(result)
 
   if (!rlang::is_dictionaryish(initial)) {
-    rlang::abort("`initial` must be a vector or list where every element has a valid name")
+    rlang::abort("Every element must have a valid and unique name in `initial`. ")
   }
 
   envir <- rlang::env(initial = as.list(initial),