+++ R/that_for.R
@@ -5,32 +5,34 @@
 #' @title R Set Comprehension
 #'
 #' @description Set comprehension with the Magrittr Pipe.
-#' Always use the syntax:
-#' `.x %>% that_for_all(.y) %>% we_have_*(f(.x, .y))`.
+#' Always use the basic syntax:
 #'
-#' @note if .y is an numeric vector, you probably want a value obtained from
-#' `peruse::range(start, end)` rather than `start:end` or `seq(start,end)`, as when
-#' start is greater than end you want an empty vector rather than counting backwards.
-#' Note that `peruse::range` views end as a supremum, not a maximum, thus `range(a,b)`
-#' is equivalent to the set `[`a,b) when a < b or `{}` when b >= a.
+#' `.x %>% that_for_all(.y) %>% we_have_*(f(.x, .y))`,
+#' but see the examples for more detail.
+#'
+#' `formula` can be anything that is recognized as a function by [purrr::as_mapper()]
+#'
+#' @note if `.y` is an numeric vector, you probably want a value obtained from
+#' `range(start, end)` rather than `start:end` or `seq.int(start,end)`,
+#' as when start is greater than end you want an empty vector rather than counting backwards.
+#' Note that [peruse::range()] views end as a supremum, not a maximum, thus `range(a,b)`
+#' is equivalent to the set `[`a,b) when a < b or the empty set when b >= a.
+#'
+#' Also note that there is some indirection in the way that `.x` and `.y` are referenced
+#' in the formula. In the function `we_have()`, the actual name of the two sets is `.x`
+#' and `.y`. That is what makes the function interface work,
+#' e.g. `function(.x, .y) .x - .y`. On the other hand, `purrr`-style lambda expressions,
+#' e.g. `~.x - .y`, use positional arguments, where `.x` is the first argument and `.y`
+#' is the second argument, no matter their names. Because those are actually their names,
+#' this difference should never matter.
+#'
 #' @param .x A set, represented as either an atomic vector or a list
 #' @param .y A set to compare to `.x`
 #' @param formula A boolean-valued function, lambda, or formula
 #' @param result Should the expression return a `vector` or an `Iterator`?
-
-#' @param that_for A list passed to `we_have`--can be ignored with proper syntax
-
-#' @details This function is designed to be used with \code{%>%} from \code{magrittr}.
-#' Essentially, this function evaluates \code{f %>% g()} as
-#' \code{eval(g(f), envir = new.env(parent = parent.frame())}. Any pipe or pipe-like
-#' function that works similarly should work with \code{that_for_*}. In particular,
-#' the R Core Team is expected to implement a base R pipe in the near future with
-#' the syntax \code{|>}. While these functions will likely work properly with the new
-#' \code{|>}, as of right now the only pipe that is guaranteed to work is
-#' \code{%>%}.
+#' @param that_for A list passed to [we_have()]—can be ignored with proper syntax
 #'
 #' @examples
-#' library(magrittr)
 #' 2:100 %>% that_for_all(range(2, .x)) %>% we_have(function(.x, .y) .x %% .y != 0) #is the same as
 #' 2:100 %>% that_for_all(range(2, .x)) %>% we_have(~.x %% .y != 0)
 #' #c.f.
@@ -40,13 +42,10 @@
 #' {c("I", "Don't", "wan't", "chicken") %>%
 #'              that_for_all("\'") %>%
 #'              we_have(~grepl(.y, .x))}
-
 #' #Twin primes 1 through 100
-#' \dontrun{
 #' primes <- 2:100 %>% that_for_all(range(2, .x)) %>% we_have(~.x %% .y != 0)
 #' primes %>% that_for_any(primes) %>% we_have(~abs(.x - .y) == 2)
 #' #Prime numbers 1 through 100 that are two away from a square number
-#' }
 #' (2:100 %>% that_for_all(range(2, .x)) %>% we_have(~.x %% .y != 0)) %>%
 #'     that_for_any(range(2, .x)) %>% we_have(~sqrt(.x + 2) == .y | sqrt(.x - 2) == .y)
 #'